\name{lab5group8-package}
\alias{lab5group8-package}
\alias{lab5group8}
\docType{package}
\title{
\packageTitle{lab5group8}
}
\description{
\packageDescription{lab5group8}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5group8}
\packageIndices{lab5group8}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab5group8}

Maintainer: \packageMaintainer{lab5group8}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }