% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{population-class}
\alias{population-class}
\alias{population}
\title{Population}
\value{
Returns an object of the class 'population'. This object can be manipulated.
}
\description{
Create an object containing population data of swedish municipalities. This function does not take any parameters but instead gets its data from the Kolada web-api.
}

